--- PCE_Web/PCE_Web/Controllers/IService.cs
@@ -1,10 +1,0 @@
-﻿using System.Collections.Generic;
-
-namespace PCE_Web.Controllers
-{
-    public interface IService<T> where T : class
-    {
-        IEnumerable<T> Get();
-    }
-
-}
