+++ PCE_Web/PCE_Web/Controllers/SearchController.cs
@@ -36,41 +36,10 @@ namespace PCE_Web.Controllers
             }
             else
             {
-                var products = GetProductsFromAPI(productName);
-                _productsCache.SetCachedItems(productName,products);
+                var products = GetProductsFromApi.GetProducts(productName);
+                _productsCache.SetCachedItems(productName, products);
                 var suggestionsView = new SuggestionsView {Products = products};
                 return View(suggestionsView);
-            }
-        }
-        //iskelti api
-        private List<Item> GetProductsFromAPI(string productName)
-        {
-            var clientHandler = new HttpClientHandler
-            {
-                ServerCertificateCustomValidationCallback = (sender, cert, chain, sslPolicyErrors) => true
-            };
-            var client = new HttpClient(clientHandler);
-            try
-            {
-                var resultList = new List<Item>();
-                var getDataTask = client.GetAsync("https://localhost:44320/api/Products/" + productName).ContinueWith(
-                    response =>
-                    {
-                        var result = response.Result;
-                        if (result.StatusCode == System.Net.HttpStatusCode.OK)
-                        {
-                            var readResult = result.Content.ReadAsAsync<List<Item>>();
-                            readResult.Wait();
-                            resultList = readResult.Result;
-                        }
-                    });
-                getDataTask.Wait();
-                return resultList;
-            }
-            catch (Exception)
-            {
-                Console.WriteLine();
-                throw;
             }
         }
     }
