+++ PCE_Web/PCE_Web/Views/Evaluation/Evaluate.cshtml
@@ -34,7 +34,7 @@
 
                     <select id="shopIdSelect" name="shopId">
                         <option value="0">--Pasirinkite--</option>
-                        @foreach (var shop in PCE_Web.Controllers.EvaluationSuccessController._allowedShopId)
+                        @foreach (var shop in PCE_Web.Controllers.EvaluationSuccessController.AllowedShopId)
                         {
                             <option value=@shop.Key>@shop.Value</option>
                         }
@@ -57,7 +57,7 @@
         <br />
         <select id="mySelect" onchange="changeIndex()">
             <option value="0">--Pasirinkite--</option>
-            @foreach (var shop in PCE_Web.Controllers.EvaluationSuccessController._allowedShopId)
+            @foreach (var shop in PCE_Web.Controllers.EvaluationSuccessController.AllowedShopId)
             {
                  <option value=@shop.Key>@shop.Value</option>
             }
@@ -65,7 +65,7 @@
         <br />
         <div class="comments-container">
             @{
-                var shopNames = Model.Comments.Join(PCE_Web.Controllers.EvaluationSuccessController._allowedShopId, c => c.ShopId, a => a.Key, (c, a) => new { a.Value });
+                var shopNames = Model.Comments.Join(PCE_Web.Controllers.EvaluationSuccessController.AllowedShopId, c => c.ShopId, a => a.Key, (c, a) => new { a.Value });
             }
             @foreach (var comment in Model.Comments)
             {
